% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{compute_cond_mean}
\alias{compute_cond_mean}
\title{A function for computing the conditional mean in multivariate normal distribution}
\usage{
compute_cond_mean(Sigma, cond.ind, X.given)
}
\arguments{
\item{Sigma}{The covariance matrix}

\item{cond.ind}{The conditional index; should lie in 1:nrow(Sigma)}

\item{X.given}{A matrix which specify the conditioned value (dim: n * length(cond.ind))}
}
\value{
A matrix of dim (nrow(Sigma) - length(cond.ind))*n
}
\description{
A function for computing the conditional mean in multivariate normal distribution
}
