% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{compute_nu}
\alias{compute_nu}
\title{Title}
\usage{
compute_nu(grid, c, B, no_nu_grid, response_type)
}
\arguments{
\item{grid}{A data frame containing different combinations of parameters}

\item{c}{A vector containing a sequence of confounding level. Default is 1:4}

\item{B}{A large number used for simulating confounding level}

\item{no_nu_grid}{The number of grid that spread from 0 to the maximum confounding
level which is searched by the algorithm}

\item{response_type}{Type of response, Gaussian or Binary}
}
\value{
A data frame containing extra two parameters: confounding level and magnitude.
}
\description{
Title
}
