% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{compute_theta_via_power}
\alias{compute_theta_via_power}
\title{A function to compute a sequence of theta}
\usage{
compute_theta_via_power(
  grid,
  maxPower,
  alpha,
  test_type,
  no_theta,
  B,
  response_type
)
}
\arguments{
\item{grid}{A parameter grid}

\item{maxPower}{The maximum power prespecified to detect maximum theta}

\item{alpha}{Level of test}

\item{test_type}{One-side or Two-side test}

\item{no_theta}{Number of theta to obtain}

\item{B}{Number of MC replications}

\item{response_type}{gaussian or binomial}
}
\value{
A new parameter grid
}
\description{
A function to compute a sequence of theta
}
