% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{create_simspec_object}
\alias{create_simspec_object}
\title{Create simulatr specifier object}
\usage{
create_simspec_object(
  alpha,
  maxType_I_null,
  Type_I_alt,
  maxPower,
  test_type,
  no_nu,
  no_theta,
  seed,
  B_nu,
  B_theta,
  B_reps,
  baseline_values,
  varying_values,
  distribution,
  way_to_learn,
  setting,
  method_strings
)
}
\arguments{
\item{alpha}{nominal level}

\item{maxType_I_null}{maximum Type-I error for naive GCM in null simulation}

\item{Type_I_alt}{Type-I error for naive GCM in alternative simulation}

\item{maxPower}{maximum power for oracle GCM in alternative simulation}

\item{test_type}{sidedness of the tests}

\item{no_nu}{number of values for nu parameter (effect of Z on Y)}

\item{no_theta}{number of values for theta parameter (effect of X on Y)}

\item{seed}{seed to control random number generation}

\item{B_nu}{number of replicates for nu computation}

\item{B_theta}{number of replicates for theta computation}

\item{B_reps}{number of repetitions for each parameter setting}

\item{baseline_values}{baseline values of each parameter}

\item{varying_values}{values along which to vary each parameter}

\item{distribution}{"binomial" or "gaussian"}

\item{way_to_learn}{"supervised" or "semi_supervised"}

\item{setting}{"null", "calibration", or "alternative"}

\item{method_strings}{list of strings specifying the methods to be applied}
}
\value{
The resulting simulatr specifier object.
}
\description{
Create simulatr specifier object
}
