% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dCRT}
\alias{dCRT}
\title{The distilled CRT test}
\usage{
dCRT(data, X_on_Z_reg, Y_on_Z_reg, hyperparams)
}
\arguments{
\item{data}{A named list with fields X, Y, Z.}

\item{X_on_Z_reg}{The regression method to apply for X|Z.}

\item{Y_on_Z_reg}{The regression method to apply for Y|Z.}

\item{hyperparams}{Additional test hyperparameters}
}
\value{
A data frame with columns "parameter," "target," "value" with p-value.
}
\description{
The distilled CRT test
}
