% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dCRT_f}
\alias{dCRT_f}
\title{Model_X test}
\usage{
dCRT_f(data, Y_on_Z_reg, resample_dist, no_resample = 2000)
}
\arguments{
\item{data}{A named list with fields X, Y, Z.}

\item{Y_on_Z_reg}{The regression method to apply for X|Z.}

\item{resample_dist}{The resampling distribution}

\item{no_resample}{The number of resamples}
}
\value{
A data frame with columns "parameter," "target," "value" with p-value.
}
\description{
Model_X test
}
