% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{E_Y_given_Z_binomial_alternative}
\alias{E_Y_given_Z_binomial_alternative}
\title{A function that computes the conditional expectation of Y|Z under alternative}
\usage{
E_Y_given_Z_binomial_alternative(theta, pred.X, pred.Y)
}
\arguments{
\item{theta}{The signal of X in Y model}

\item{pred.X}{Predictor of X given Z}

\item{pred.Y}{Predictor of Y given Z}
}
\value{
A positive value.
}
\description{
A function that computes the conditional expectation of Y|Z under alternative
}
