% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{fast_data_generate}
\alias{fast_data_generate}
\title{A fast way to generate data from multivariate Gaussian distribution}
\usage{
fast_data_generate(mean, covariance, B, n)
}
\arguments{
\item{mean}{The mean vector of MVN.}

\item{covariance}{The covariance matrix of MVN.}

\item{B}{The number of replicate}

\item{n}{The number of sample}
}
\value{
A data matrix generated via Cholesky
}
\description{
A fast way to generate data from multivariate Gaussian distribution
}
