% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{fit_conditional_variance}
\alias{fit_conditional_variance}
\title{Title}
\usage{
fit_conditional_variance(response, features, conditional_mean, method)
}
\arguments{
\item{response}{A response vector of length n}

\item{features}{A design matrix of dimension nxp}

\item{conditional_mean}{The conditional mean vector of length n}

\item{method}{A string that specifies the variance estimation method to use, e.g. \code{squared_residual}}
}
\value{
A vector of estimated conditional variances of length n
}
\description{
Title
}
