% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GCM_debug}
\alias{GCM_debug}
\title{A function that outputs the bias term in GCM test}
\usage{
GCM_debug(data, X_on_Z_reg, Y_on_Z_reg, test_hyperparams)
}
\arguments{
\item{data}{A named list with fields X, Y, Z.}

\item{X_on_Z_reg}{The regression method to apply for X|Z.}

\item{Y_on_Z_reg}{The regression method to apply for Y|Z.}

\item{test_hyperparams}{Additional method hyperparameters (currently not used)}
}
\value{
A data frame with columns "parameter," "target," "value",
with three rows, are respectively cross_bias, X_Z_bias, Y_Z_bias
}
\description{
A function that outputs the bias term in GCM test
}
