% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GCM_f}
\alias{GCM_f}
\title{The generalized covariance measure test of Shah and Peters.}
\usage{
GCM_f(data, X_on_Z_reg, Y_on_Z_reg, X_on_Z_var)
}
\arguments{
\item{data}{A named list with fields X, Y, Z.}

\item{X_on_Z_reg}{The regression method to apply for X|Z.}

\item{Y_on_Z_reg}{The regression method to apply for Y|Z.}

\item{X_on_Z_var}{The method to obtain the conditional variance of X|Z.}
}
\value{
A data frame with columns "parameter," "target," "value",
with two rows, one for the test statistic, and one for the p-value.
}
\description{
\code{GCM_f} is a function carrying out the dCRT.
}
