% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_cov_ar1}
\alias{generate_cov_ar1}
\title{Generate dxd AR(1) covariance matrix with parameter rho}
\usage{
generate_cov_ar1(rho, d)
}
\arguments{
\item{d}{}
}
\value{
A dxd covariance matrix.
}
\description{
Generate dxd AR(1) covariance matrix with parameter rho
}
