% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_data}
\alias{generate_data}
\title{Generic function to generate data for the numerical simulations}
\usage{
generate_data(
  n,
  N,
  d,
  rho,
  B,
  coef_neg,
  coef_pos,
  nu,
  theta,
  distribution,
  setting
)
}
\arguments{
\item{n}{Sample size}

\item{d}{Dimension of Z}

\item{rho}{Autocorrelation parameter for Z}

\item{B}{Number of samples to draw}

\item{coef_neg}{List of negative coefficient indices}

\item{coef_pos}{List of positive coefficient indices}

\item{nu}{Effect size of Z on Y}

\item{theta}{Effect size of X on Y}

\item{distribution}{Either "binomial" or "gaussian"}

\item{setting}{Either "null", "calibration", or "alternative"}
}
\value{
A list of B datasets
}
\description{
Generic function to generate data for the numerical simulations
}
