% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_glm_response_data}
\alias{generate_glm_response_data}
\title{Generate response data based on a GLM model}
\usage{
generate_glm_response_data(design_matrix, coefficients, family)
}
\arguments{
\item{design_matrix}{The design matrix (n x p)}

\item{coefficients}{Coefficient vector (vector of length p)}

\item{family}{One of \code{gaussian, binomial, poisson}}
}
\value{
A response vector of length n
}
\description{
Generate response data based on a GLM model
}
