% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_method_df_from_strings}
\alias{generate_method_df_from_strings}
\title{Helper function to translate a list of method specifications strings into a
data frame for use with \code{generate_method_list_from_df}}
\usage{
generate_method_df_from_strings(method_strings, distribution, way_to_learn)
}
\arguments{
\item{method_strings}{A vector of strings of the kind we define in create_simspec_objects.R}

\item{distribution}{Either "binomial" or "gaussian"}

\item{way_to_learn}{Either "supervised" or "semi_supervised"}
}
\value{
A data frame as expected by \code{generate_method_list_from_df}
}
\description{
Helper function to translate a list of method specifications strings into a
data frame for use with \code{generate_method_list_from_df}
}
