% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_method_list}
\alias{generate_method_list}
\title{Helper function to translate the data frame of method specifications into a
list of simulatr functions.}
\usage{
generate_method_list(methods_df)
}
\arguments{
\item{methods_df}{A data frame of the kind we define in sim_setting_x.R}
}
\value{
A list of simulatr functions whose length is the number of rows of \code{methods_df}.
}
\description{
Helper function to translate the data frame of method specifications into a
list of simulatr functions.
}
