% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{generate_method_list_from_strings}
\alias{generate_method_list_from_strings}
\title{Helper function to translate the data frame of method specifications into a
list of simulatr functions.}
\usage{
generate_method_list_from_strings(method_strings, distribution, way_to_learn)
}
\arguments{
\item{method_strings}{A vector of strings of the kind we define in create_simspec_objects.R}

\item{distribution}{Either "binomial" or "gaussian"}

\item{way_to_learn}{Either "supervised" or "semi_supervised"}
}
\value{
A list of simulatr functions whose length is same as that of \code{method_strings}.
}
\description{
Helper function to translate the data frame of method specifications into a
list of simulatr functions.
}
