% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{magnitude_detect}
\alias{magnitude_detect}
\title{A line search algorithm for detecting the magnitude of coefficient reaching the confounding level}
\usage{
magnitude_detect(
  n,
  data,
  c,
  alpha,
  beta,
  gamma,
  eps = 1e-04,
  type = "gaussian"
)
}
\arguments{
\item{n}{The number of data}

\item{data}{A data list containing the noise vector for X|Z and Y|Z and data matrix Z}

\item{c}{The target confounding level}

\item{alpha}{The step size when doing line search}

\item{beta}{The coefficient vector for Y|Z}

\item{gamma}{The coefficient vector for X|Z}

\item{eps}{The precision number}

\item{type}{The type of response (either gaussian or binomial)}
}
\value{
The magnitude of coefficient reaches the confounding level
}
\description{
A line search algorithm for detecting the magnitude of coefficient reaching the confounding level
}
