% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{MSE_shared}
\alias{MSE_shared}
\title{This is a function for computing MSE on shared variables}
\usage{
MSE_shared(data, coef_hat, type, cond_distr)
}
\arguments{
\item{data}{A list containing oracle coefficient, data, number of nonzero component etc.}

\item{coef_hat}{A vector containing the estimated coefficient}

\item{type}{Either under null or alternative}

\item{cond_distr}{Conditional distribution including X|Z and Y|Z}
}
\value{
A positive number indicating the MSE on shared variables
}
\description{
This is a function for computing MSE on shared variables
}
