% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{MSE_total}
\alias{MSE_total}
\title{This is a function for computing MSE on total variables}
\usage{
MSE_total(oracle_pred, covariate, coef_hat)
}
\arguments{
\item{oracle_pred}{The oracle conditional expectation}

\item{covariate}{Z matrix}

\item{coef_hat}{A vector containing the estimated coefficient}
}
\value{
A positive number indicating the MSE on total variables
}
\description{
This is a function for computing MSE on total variables
}
