% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{MX2_F_test_f}
\alias{MX2_F_test_f}
\title{The MX(2) F-test.}
\usage{
MX2_F_test_f(data, X_on_Z_reg, Y_on_Z_reg)
}
\arguments{
\item{data}{A named list with fields X, Y, Z.}

\item{X_on_Z_reg}{The regression method to apply for X|Z.}

\item{Y_on_Z_reg}{The regression method to apply for Y|Z.}

\item{X_on_Z_var}{The method to obtain the conditional variance of X|Z.}
}
\value{
A data frame with columns "parameter," "target," "value",
with two rows, one for the test statistic, and one for the p-value.
}
\description{
\code{MX2_F_test_f} is a function carrying out the MX(2) F-test.
}
