% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Title}
\usage{
orthogonalize(gZ, Z_sub)
}
\arguments{
\item{gZ}{The predictor vector}

\item{Z_sub}{The submatrix selected by user for g(Z)}
}
\value{
A converted, stacked g(Z)
}
\description{
Title
}
