% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{pval_shift}
\alias{pval_shift}
\title{Compute the shifted p_value}
\usage{
pval_shift(alpha, c, type = "two_side")
}
\arguments{
\item{alpha}{Level of test}

\item{c}{Confounding level}

\item{type}{Type of test: one-sided or two-sided}
}
\value{
Shifted p_value
}
\description{
Compute the shifted p_value
}
