% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{resample_dCRT}
\alias{resample_dCRT}
\title{resampling function to get resample with fitted mean and variance}
\usage{
resample_dCRT(
  conditional_mean,
  conditional_variance = NULL,
  no_resample = 1000,
  resample_dist
)
}
\arguments{
\item{conditional_mean}{A vector of result from \code{fit_conditional_mean}}

\item{conditional_variance}{A vector of result from \code{fit_conditional_variance}}

\item{no_resample}{A number that specifies the size of resamples}

\item{resample_dist}{A distribution that specifies the resampling distribution \code{Binom} or \code{Gaussian}}
}
\value{
A matrix of nxno_resample including the resamples
}
\description{
resampling function to get resample with fitted mean and variance
}
