% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{set_default_method_hyperparams}
\alias{set_default_method_hyperparams}
\title{Populate default hyperparameters for each method}
\usage{
set_default_method_hyperparams(method_type, hyperparams)
}
\arguments{
\item{method_type}{The method type (MLE, LASSO, etc.)}

\item{hyperparams}{The named list of hyperparams}
}
\value{
The same named list that was input, with default values added if necessary
}
\description{
Populate default hyperparameters for each method
}
