% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simulate_confounding}
\alias{simulate_confounding}
\title{A function for calculating the confoudning level via simulation}
\usage{
simulate_confounding(n, X, Y)
}
\arguments{
\item{n}{number of samples}

\item{X}{A vector of treatment (length may be much larger than n)}

\item{Y}{A vector of outcome (length may be much larger than n)}
}
\value{
The confounding level
}
\description{
A function for calculating the confoudning level via simulation
}
