% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simulate_power}
\alias{simulate_power}
\title{A function for calculating the approximation power for oracle GCM}
\usage{
simulate_power(n, X, Y, E_X_given_Z, E_Y_given_Z)
}
\arguments{
\item{n}{number of samples}

\item{X}{A vector of treatment (length may be much larger than n)}

\item{Y}{A vector of outcome (length may be much larger than n)}

\item{E_X_given_Z}{A vector of conditional expectation of X|Z}

\item{E_Y_given_Z}{A vector of conditional expectation of Y|Z}
}
\value{
A number representing the power
}
\description{
A function for calculating the approximation power for oracle GCM
}
