% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symcrt-package.R
\docType{package}
\name{symcrt-package}
\alias{symcrt}
\alias{symcrt-package}
\title{symcrt: A collection of methods for the symcrt project}
\description{
Implementations of various methods for conditional independence testing, including the MX(2) F-test, dCRT, and GCM.
}
\author{
\strong{Maintainer}: Eugene Katsevich \email{ekatsevi@wharton.upenn.edu}

Authors:
\itemize{
  \item Ziang Niu \email{ziangniu@upenn.edu}
}

}
\keyword{internal}
