% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dCRT_f}
\alias{dCRT_f}
\title{The distilled CRT test}
\usage{
dCRT_f(
  data,
  X_on_Z_reg,
  Y_on_Z_reg,
  X_on_Z_var,
  resample_dist,
  no_resample = 2000
)
}
\arguments{
\item{data}{A named list with fields X, Y, Z.}

\item{X_on_Z_reg}{The regression method to apply for X|Z.}

\item{Y_on_Z_reg}{The regression method to apply for Y|Z.}

\item{X_on_Z_var}{The method to obtain the conditional variance of X|Z.}

\item{resample_dist}{The resampling distribution for CRT}

\item{no_resample}{The number of resampling}
}
\value{
A data frame with columns "parameter," "target," "value" with p-value.
}
\description{
The distilled CRT test
}
