% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{fast_data_generate}
\alias{fast_data_generate}
\title{Title}
\usage{
fast_data_generate(covariance, B, n)
}
\arguments{
\item{covariance}{The covariance matrix of MVN.}

\item{B}{The number of replicate}

\item{n}{The number of sample}
}
\value{
A data matrix generated via Cholesky
}
\description{
Title
}
