% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_helpers.R
\name{fit_conditional_mean}
\alias{fit_conditional_mean}
\title{Fit conditional mean.}
\usage{
fit_conditional_mean(response, features, method)
}
\arguments{
\item{response}{A response vector of length n}

\item{features}{A design matrix of dimension nxp}

\item{method}{A string that specifies which regression method to use, e.g. \code{OLS} or \code{LASSO}}
}
\value{
A vector of fitted conditional means
}
\description{
Function that fits some regression method to get a conditional mean of a
response (could be Y or X) given features (usually Z).
}
