% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simulate_confounding}
\alias{simulate_confounding}
\title{A function for calculating the confoudning level via simulation}
\usage{
simulate_confounding(X, Y)
}
\arguments{
\item{X}{A vector of treatment}

\item{Y}{A vector of outcome}
}
\value{
The confounding level
}
\description{
A function for calculating the confoudning level via simulation
}
